Multi-platform builds on GH actions
Updates the postgrest-rs dependency to the latest, which was just
rebased on upstream. The upstream postgrest-rs project recently moved
to using rustls exclusively, and this commit also changes our reqwest
dependency to use rustls, so that there's only a single tls
implementation in use. The end result is that we can now build a
statically linked binary on linux using musl.

This also adds github actions workflows for CI and publishing releases.
The intent is that we manually create releases using the UI, and the
actions will run to build and publish the release binaries. This
includes builds for both macos (x86 and arm) and linux (x86 only).

The mac build is produced as a single "fat" binary that works on both
intel and m1 macs.